% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mp_gmm.R
\name{generate_mp_gmm}
\alias{generate_mp_gmm}
\title{Generate multi-profile Gaussian mixture data}
\usage{
generate_mp_gmm(
  P,
  L_vec,
  K_vec,
  feature_group_proportions,
  mixing_proportions,
  dist_mahalanobis,
  target_ari_features,
  M,
  N,
  covariance_spec,
  add_noise = FALSE,
  noise_sd = 0.1,
  seed = NULL,
  noise_feature_fraction = 0,
  ari_mode = "vs_reference",
  ari_tol = 0.01,
  ari_max_iter = 10000,
  ari_swap_frac = 0.1,
  profile_weights = NULL,
  template_sd = 1,
  feature_sd_within_group = 0,
  baseline_sd = 0.5
)
}
\arguments{
\item{P}{Number of profiles.}

\item{L_vec}{Number of feature groups in each profile.}

\item{K_vec}{Number of observation components in each profile.}

\item{feature_group_proportions}{List of feature-group proportions.}

\item{mixing_proportions}{List of component proportions.}

\item{dist_mahalanobis}{Target pairwise Mahalanobis separation.}

\item{target_ari_features}{ARI control for feature partitions.}

\item{M}{Number of features.}

\item{N}{Number of observations.}

\item{covariance_spec}{Covariance specification.}

\item{add_noise}{Should additive Gaussian noise be added?}

\item{noise_sd}{Standard deviation of additive noise.}

\item{seed}{Optional random seed.}

\item{noise_feature_fraction}{Fraction of features replaced by pure noise.}

\item{ari_mode}{ARI control mode.}

\item{ari_tol}{Tolerance for ARI optimization.}

\item{ari_max_iter}{Maximum number of ARI optimization iterations.}

\item{ari_swap_frac}{Fraction of labels swapped in ARI perturbation.}

\item{profile_weights}{Optional profile weights.}

\item{template_sd}{Standard deviation of group-level loadings.}

\item{feature_sd_within_group}{Feature-level deviation within a group.}

\item{baseline_sd}{Standard deviation of feature baselines.}
}
\value{
An object of class \code{"mp_gmm_data"}.
}
\description{
Generate a synthetic data matrix with multiple latent profiles.
}
