% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.MPObject}
\alias{print.MPObject}
\title{Print a multi-profile GMM object}
\usage{
\method{print}{MPObject}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"MPObject"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The input object, invisibly.
}
\description{
Print a multi-profile GMM object
}
