% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.multprofile_gmm_data}
\alias{print.multprofile_gmm_data}
\title{Print a multi-profile GMM object}
\usage{
\method{print}{multprofile_gmm_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"multprofile_gmm_data"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The input object, invisibly.
}
\description{
Print a multi-profile GMM object
}
