% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.MPObject}
\alias{summary.MPObject}
\title{Summarize a multi-profile GMM object}
\usage{
\method{summary}{MPObject}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"MPObject"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The input object, invisibly.
}
\description{
Summarize a multi-profile GMM object
}
