% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.multprofile_gmm_data}
\alias{summary.multprofile_gmm_data}
\title{Summarize a multi-profile GMM object}
\usage{
\method{summary}{multprofile_gmm_data}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"multprofile_gmm_data"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The input object, invisibly.
}
\description{
Summarize a multi-profile GMM object
}
